clean up code

+++ app/controllers/transactions_controller.rb
@@ -52,9 +52,8 @@ class TransactionsController < ApplicationController
         @transaction.destroy 
 
         #if user has no transactions with coin_id, delete wallet
-        if current_user.transactions.find_by(coin_id: coin_id).nil?
-            wallet.destroy
-        end
+        wallet.destroy if current_user.transactions.find_by(coin_id: coin_id).nil?
+        
         redirect_to user_transactions_path(current_user), notice: "Transaction has been deleted!"
     end
